Moving useEffect Gradient to GameCanvas
