+++ frontend/src/GameComponents/GamePage/GamePage.js
@@ -84,23 +84,6 @@ function GamePage() {
     setDifficulty(newDifficulty);
   };
 
-  useEffect(() => {
-    const canvas = document.getElementById("GameCanvas");
-    if (canvas) {
-        var c = document.getElementById("GameCanvas");
-        var ctx = c.getContext("2d");
-
-        // Create gradient
-        var grd = ctx.createLinearGradient(0, 0, 200, 0);
-        grd.addColorStop(0, "red");
-        grd.addColorStop(1, "white");
-
-        // Fill with gradient
-        ctx.fillStyle = grd;
-        ctx.fillRect(10, 10, 150, 80);
-      } 
-  }, []);
-
   // Once user has answered all questions, display score and continue button
   if (gameOver) {
     return (
